pyproject.toml
[project]
name = "cs117_asv"
version = "0.1.0"
description = "cs117 asv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "comet-ml>=3.49.11",
    "librosa>=0.11.0",
    "numpy>=2.2.6",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "soundfile>=0.13.1",
    "tensorboard>=2.19.0",
    "torch>=2.7.1",
    "torchcontrib>=0.0.2",
    "tqdm>=4.67.1",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
]
